Rename mapController.js -> js.erb for musician asset pathing
Access to the assets is defined in different
directories depending on the Rails environment.
Hard-coding the path as `assets/img-name.png` only
works in the development mode.  When we go to
production mode (e.g. Heroku), this will not work
and we will not be able to see our sweet custom
asset pins.

Fix this by using the `asset_path` helper in Rails.

This requires that something interpolate ERB into
JavaScript.  Ergo, we add the file suffix `.erb`
so that the Asset Pipeline will turn this into
interpolated JavaScript with the correct path.

For further reference see:

http://guides.rubyonrails.org/asset_pipeline.html